Tests/LibWeb/Text/expected/css/placeholder-letter-spacing-zero.txt

input placeholder 'letter-spacing: 0': normal
